Code Refactoring
Eliminate le classi per gestire gli AlertBox. Ora c'è la classe AlertBoxManager che gestisce i vari alertbox
